Fix import statements for moved components (Part 2)
- Updated imports in ResetPassword.tsx and SignUp.tsx to use shared components
- Fixed imports in UserDashboard.tsx to use feature-based imports and shared components
- Corrected relative paths to match new folder structure
- Fixed function name typo in SignUp.tsx (signUpAuthentication)
