+++ src/features/user/pages/UserDashboard/UserDashboard.tsx
@@ -1,12 +1,12 @@
 import { useEffect } from "react";
-
-import UsersFantasyRanking from "../../components/UsersFantasyRanking/UsersFantasyRanking";
-import UsersFortyRanking from "../../components/UsersFortyRanking/UsersFortyRanking";
-import UsersResults from "../../components/UsersResults/UsersResults";
-import UsersUpcomingBattles from "../../components/UsersUpcomingBattles/UsersUpcomingBattles";
+import { UsersFantasyRanking, UsersFortyRanking, UsersResults, UsersUpcomingBattles } from "../../index";
 import "./UserDashboard.scss";
 import { useNavigate } from "react-router-dom";
 import { CircularProgress } from "@mui/material";
+import { useUserStore } from "../../../store/user";
+import { useBattlesStore } from "../../../store/battles";
+import { useArmiesStore } from "../../../store/armies";
+import { DashboardHero } from "../../../shared";
 
 export default function UserDashboard() {
   const { token, userInfo, fetchUserInfo } = useUserStore();
